--- app/src/main/java/com/hongsup/explog/view/postdetail/PostDetailActivity.java
@@ -1,26 +1,0 @@
-package com.hongsup.explog.view.postdetail;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-
-import com.hongsup.explog.R;
-
-import butterknife.ButterKnife;
-
-public class PostDetailActivity extends AppCompatActivity {
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_post_detail);
-        ButterKnife.bind(this);
-
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        getSupportActionBar().setTitle("");
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-    }
-}
