+++ app/src/main/res/drawable/border_cover_tag.xml
@@ -1,0 +1,8 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <solid android:color="#4A99D5" />
+    <corners android:radius="30dp" />
+    <size
+        android:width="24dp"
+        android:height="24dp" />
+</shape>
