+++ app/src/main/res/menu/menu_post.xml
@@ -1,11 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
+
     <item
-        android:id="@+id/action_ok"
-        android:title="OK"
-        android:icon="@drawable/ic_check_white_24dp"
-        app:showAsAction="always"
+        android:id="@+id/action_cover_change"
+        android:title="@string/menu_post_cover_change"
+        app:showAsAction="withText"
+        />
+
+    <item
+        android:id="@+id/action_delete"
+        android:title="@string/menu_post_delete"
+        app:showAsAction="withText"
         />
 
 </menu>
